services/model-runtime/pyproject.toml
[project]
name = "model-runtime"
version = "0.1.0"
description = "Strict local inference /infer/* + /solve/mechanics (engineering-core)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.27",
  "pydantic>=2",
  "pyyaml>=6",
  "jsonschema>=4.20",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
hf = [
  "torch>=2.0.0",
  "transformers>=4.39",
  "accelerate>=0.30.0",
]

[tool.pytest.ini_options]
addopts = [
  "-o",
  "cache_dir=../../.cache/pytest/services-model-runtime",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
